+++ OpenGL-basico/utils/particles_handler.h
@@ -11,9 +11,9 @@ private:
     std::vector<particle*> particles_;
 public:
     static particles_handler* get_instance();
-    void update(int seconds);
-    void add(int seconds, vector3 position, vector3 velocity, vector4 initial_color, vector4 final_color, int size, int life_time);
+    void update(float seconds);
+    void add(float seconds, vector3 position, vector3 velocity, vector4 initial_color, vector4 final_color, float size, float life_time);
     std::vector<particle*> get_particles();//USAR DESPUES DE UPDATE!!!
-    void create_explotion(int seconds, vector3 position, int number_of_particles);
-    void walk_particles(int seconds, vector3 position, vector3 speed, int number_of_particles);
+    void create_explotion(float seconds, vector3 position, int number_of_particles);
+    void walk_particles(float seconds, vector3 position, vector3 speed, int number_of_particles);
 };
